Add files via upload
In C++ accessing an array out of bounds does not show any error and it might cause problems in computations by showing segmentation fault. Moreover, since each time the main code ( which is calling these arrays) is accessing a location which does not exist leading to memory issues and bugs. These updated header files will abort if an array is accessed out of bounds.